Split YamlReadingTest into several files
